+++ src/main/java/duke/Duke.java
@@ -1,3 +1,5 @@
+package duke;
+import duke.command.Command;
 public class Duke {
     private final Storage storage;
     private final TaskList taskList;
