+++ src/main/java/duke/Parser.java
@@ -1,3 +1,16 @@
+package duke;
+
+import duke.command.Command;
+import duke.command.DeadlineCommand;
+import duke.command.DeleteCommand;
+import duke.command.EventCommand;
+import duke.command.ExitCommand;
+import duke.command.ListCommand;
+import duke.command.MarkCommand;
+import duke.command.ToDoCommand;
+import duke.command.UnmarkCommand;
+import duke.command.WrongCommand;
+
 public class Parser {
     public static Command parse(String message) {
         String firstWord = message.split(" ")[0].toLowerCase();
