+++ src/main/java/duke/Storage.java
@@ -1,3 +1,6 @@
+package duke;
+
+import duke.task.Task;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
