+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,9 @@
+package duke;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
 import java.time.LocalDate;
 import java.util.ArrayList;
