+++ src/main/java/duke/command/EventCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
