+++ src/main/java/duke/task/Task.java
@@ -1,3 +1,4 @@
+package duke.task;
 import java.io.FileWriter;
 import java.io.IOException;
